:minibus: Reorganize command helpers
