+++ src/command/helpers/figures.js
