+++ src/command/helpers/index.js
@@ -1,0 +1,19 @@
+// @flow
+
+import tildify from 'tildify'
+import Color from 'cli-color'
+import Figure from './figures'
+import Table from './table'
+import Symbol from './symbols'
+import prompt from './prompt'
+import gitStatus from './git-status'
+
+export default {
+  tildify,
+  Color,
+  Figure,
+  Table,
+  Symbol,
+  prompt,
+  gitStatus,
+}
