+++ src/command/helpers/symbols.js
