+++ src/commands/EjectCommand.js
@@ -13,7 +13,7 @@ export default class EjectCommand extends Command {
 
   async run({ config }: Object, ...list: Array<string>) {
     const directories = list || ['.']
-    const { Color, Figure, tildify: tld } = this.utils
+    const { Color, Figure, tildify: tld } = this.helpers
 
     await this.ensureProjectsRoot()
     const ejects = []
@@ -33,7 +33,7 @@ export default class EjectCommand extends Command {
     this.log(`${directories.join(' ')}`)
   }
   async eject(config: string, directory: string) {
-    const { Color, tildify, prompt } = this.utils
+    const { Color, tildify, prompt } = this.helpers
 
     const sourceDir = Path.resolve(directory)
     const sourceDirList = sourceDir.split(Path.sep)
