--- src/context-utils/index.js
@@ -1,6 +1,0 @@
-export Figure from './figures'
-export Table from './table'
-export Symbol from './symbols'
-export prompt from './prompt'
-export Color from 'cli-color'
-export tildify from 'tildify'
