+++ src/helpers.js
@@ -2,11 +2,9 @@
 
 import invariant from 'assert'
 import expandTilde from 'expand-tilde'
-
 import type { Options, ParsedRepo } from './types'
 
 export const CONFIG_FILE_NAME = '.repoman.json'
-
 export const BUILTIN_COMMANDS = new Set([
   'get',
   'get-config',
@@ -26,9 +24,6 @@ export class RepoManError extends Error {
   }
 }
 
-export function processPath(path: string): string {
-  return expandTilde(path)
-}
 // Spec:
 // If one is given, validate it to be a valid string
 // Return the expanded tilde
@@ -47,7 +42,7 @@ export function fillConfig(given: Object): Options {
   } else {
     options.stateDirectory = '~/.repoman'
   }
-  options.stateDirectory = processPath(options.stateDirectory)
+  options.stateDirectory = expandTilde(options.stateDirectory)
 
   return options
 }
